+++ src/pages/posts/vdom-compilers/index.mdx
@@ -1,4 +1,4 @@
-import StateEquality from '../../components/vdom-compilers/State-Equality.tsx'
+import StateEquality from './State-Equality.tsx'
 
 export const documentProps = {
   title: 'Virtual Dom and compilers',
@@ -8,8 +8,9 @@ export const documentProps = {
 
 # Virtual Dom and compilers
 
-As one of the core maintainers of Preact I think a lot about VDOM and how to optimize, there are a lot of
-optmizations that are easy to miss in Virtual DOM. We often get told to leverage `shouldComponentUpdate` or `memo`.
+As one of the core maintainers of Preact I think a lot about Virtual DOM and how to optimize it, there are a lot of
+optmizations that are easy to miss in Virtual DOM. We often get told to leverage `shouldComponentUpdate` or `memo` to
+avoid excessive rerendering of a component.
 These are a few of the more obvious optmizations as they are explicitly documented and we can reaon about them while
 we explicitly use them. Let's look at a few of the less obvious ones, in the examples we'll notice that a rerender of
 a component will switch the border-color for a bit so we can see how many times everything rerenders.
@@ -20,6 +21,8 @@ When we look at the example of `setState` we have a few advantages with using si
 opposed to an object or array. It's much easier for a library to assess whether these are equal
 
 <StateEquality />
+
+In the big picture this looks obvious but in some situations this can lead to some unexpected cases
 
 ## Strict equality
